refactor: clean up imports in authentication routes
