+++ server/src/shared_kernel/authentication/__init__.py
@@ -1,7 +1,4 @@
 from .models import ValidatedUser, AuthenticatedUser
-from .dependencies import (
-    get_current_user,
-)
 from .exceptions import (
     AuthenticationError,
     InvalidTokenError,
@@ -16,7 +13,6 @@ from .constants import ADMIN_ROLE
 __all__ = [
     "ValidatedUser",
     "AuthenticatedUser",
-    "get_current_user",
     "AuthenticationError",
     "InvalidTokenError",
     "InsufficientPermissionsError",
