+++ server/src/vault_management_context/adapters/primary/fastapi/routes/vault_unlock_routes.py
@@ -10,7 +10,8 @@ from vault_management_context.application.use_cases.unlock_vault_use_case import
 )
 from vault_management_context.domain.entities.share import Share
 from vault_management_context.domain.exceptions import VaultManagementDomainError
-from shared_kernel.authentication import get_current_user, ValidatedUser, NotAdminError
+from shared_kernel.authentication import ValidatedUser, NotAdminError
+from shared_kernel.authentication.dependencies import get_current_user
 
 router = APIRouter(prefix="/vault", tags=["Vault"])
